requirements.txt
litellm>=1.30.3
python-dotenv>=1.0.0
pyyaml>=6.0.0